+++ src/cpp/stdafx.h
@@ -130,12 +130,13 @@ using FullIdx = int64_t;
 /**
  * \brief Type used for perimeter raster values (uses [0, 1])
  */
-// FIX: change to use smaller type at some point
-// using PerimSize = uint8_t;
-using PerimSize = uint16_t;
+// FIX: could try to get really fancy and use number of bits per pixel options
+using PerimSize = uint8_t;
+// using PerimSize = uint16_t;
 /**
  * \brief Type used for fuel values (uses [0 - 999]?)
  */
+// FIX: seriously does not like uint for some reason
 using FuelSize = uint16_t;
 // using FuelSize = int16_t;
 /**
